+++ src/extension.ts
@@ -5,11 +5,12 @@ import { execSync } from 'child_process';
 import * as vscode from 'vscode';
 import { Observable } from 'rxjs';
 
-import { CreateConfig } from './config.interface';
+import { Config } from './config.interface';
+import { getConfigPath } from './utils';
 
 export function activate(context: vscode.ExtensionContext) {
-    let genBEMFile = vscode.commands.registerCommand('extension.newBEMFile', (uri: vscode.Uri) => {
-        let createConfig: CreateConfig = <CreateConfig>vscode.workspace.getConfiguration('bemy.create');
+    let bemyCreate = vscode.commands.registerCommand('extension.bemyCreate', (uri: vscode.Uri) => {
+        let config: Config = <Config>vscode.workspace.getConfiguration('bemy');
 
         let templateInputDialog = Observable.from(
             vscode.window.showInputBox(
@@ -24,9 +25,11 @@ export function activate(context: vscode.ExtensionContext) {
                 }
             })
             .do(type => {
-                let addToGit: string = createConfig.addToGit ? '-g' : '';
+                let addToGit: string = config.addToGit ? '-g' : '';
+                let bemyConfigPath: string = getConfigPath(config);
+                let configArg: string = bemyConfigPath ? `-c ${bemyConfigPath}` : '';
 
-                const command = `bemy -t create -f ${uri.fsPath} -p "${type}" ${addToGit}`;
+                const command = `bemy -t create -f ${uri.fsPath} -p "${type}" ${addToGit} ${configArg}`;
                 execSync(command, { cwd: vscode.workspace.rootPath });
             })
             .subscribe(
@@ -35,7 +38,53 @@ export function activate(context: vscode.ExtensionContext) {
             );
     });
 
-    context.subscriptions.push(genBEMFile);
+    let bemyRename = vscode.commands.registerCommand('extension.bemyRename', (uri: vscode.Uri) => {
+        let config: Config = <Config>vscode.workspace.getConfiguration('bemy');
+
+        let templateInputDialog = Observable.from(
+            vscode.window.showInputBox(
+                { prompt: 'Please enter new name' }
+            )
+        );
+
+        templateInputDialog
+            .do(value => {
+                if (!value) {
+                    throw new Error('You should enter new name');
+                }
+            })
+            .do(newName => {
+                let addToGit: string = config.addToGit ? '-g' : '';
+                let bemyConfigPath: string = getConfigPath(config);
+                let configArg: string = bemyConfigPath ? `-c ${bemyConfigPath}` : '';
+                let deepRename: string = config.deepRename ? '-d' : '';
+
+                const command = `bemy -t rename -f ${uri.fsPath} -p "${newName}" ${addToGit} ${deepRename} ${configArg}`;
+                execSync(command, { cwd: vscode.workspace.rootPath });
+            })
+            .subscribe(
+                () => vscode.window.setStatusBarMessage('Files renamed', 5000),
+                err => vscode.window.showErrorMessage(err.message)
+            );
+    });
+
+    let bemyDefault = vscode.commands.registerCommand('extension.bemyDefault', (uri: vscode.Uri) => {
+
+        const command = `bemy -f ${uri.fsPath}`;
+
+        Observable
+            .from(
+                execSync(command, { cwd: vscode.workspace.rootPath })
+            )
+            .subscribe(
+                () => vscode.window.setStatusBarMessage('Task executed', 5000),
+                err => vscode.window.showErrorMessage(err.message)
+            );
+    });
+
+    context.subscriptions.push(bemyCreate);
+    context.subscriptions.push(bemyRename);
+    context.subscriptions.push(bemyDefault);
 }
 
 export function deactivate() {}