+++ CHANGELOG.md
@@ -6,3 +6,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.0]
 ### Added
 - call bemy create from folder context menu in explorer
+
+## [1.1.0]
+### Added
+- adds `rename` and `default` task
+- adds Bemy tasks from context menu
+- adds configurable cli task
+- adds key bindings